extracted class GameSizingComputer
