+++ gazeplay/src/main/java/net/gazeplay/games/whereisit/GameSizing.java
@@ -1,0 +1,12 @@
+package net.gazeplay.games.whereisit;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+class GameSizing {
+    public final double width;
+    public final double height;
+    public final double shift;
+}
